erickshaw/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
